Milestone 2-5
Updated BTT for NPC movement so its speed will depend on how many keys the player has collected (code in Lock & Key BP increments a value in the BB). Changed the nav mesh to be dynamic so the NPCs will run through a part of the maze that previously had a door. Added the NPC's ability to teleport the player back to the start of the maze when they catch the player (I make sure they immediately forget the player too). Tried adding sprint/crouch options for player in C++, unsuccessfully.