Milestone 2-7
Attempting giving the player health UI, couldn't get it to work. Gave the player current health and max health, and functions to take damage and die. Instead of locking the player's controls, dying will reset their current health and teleport them to the beginning of the maze, a vector set by designers. Some things were exposed to BP so the AI could deal damage to the player and check the player's current health to determine if they were just teleported and immediately forget the player.