Im Sad
Added the maze save game c++ class and gave it 4 variables for the things we want to save, but I can't do any more than that right now. Can't interact with something that's not actually in the project cause none of the c++ classes will compile until I get back home.