Audio Saved
Now that I can compile the C++ code, I added the victory trigger boxes to the levels, tested all the UI, and added the code in BP to save the audio volume slider positions using the MazeSaveGame class, but the variable for the stage beaten is not used yet. For some reason Unreal has been running slow and crashing relentlessly, though inconsistently. I thought I figured out the problem came down to the node Set Focus to Game Viewport, but then Unreal continues to usually crash before the death/victory screen appears, and sometimes it doesn't.