Bonus Features
Added an exit sign for the areas behind the final door, next to the victory trigger box. Added a widget for the player to see the controls in the first level. Gave the player a new ability to ping visible enemies and enabling the player to see pinged enemies behind walls for 40 seconds.