Improved Ping
Unfortunately the bug with moving back right while crouched seems to be a problem with the engine, and the only partial solution was to make the crouch a toggle so that as long as the player isn't touching C, they can move back right. The player's ping ability has also been improved to guarantee ping all visible enemies in the event they are lined up, one enemy can't block the line trace to another.